+++ edison-jobs/src/main/java/de/otto/edison/jobs/repository/JobLockRepository.java
@@ -1,15 +1,11 @@
 package de.otto.edison.jobs.repository;
 
-import de.otto.edison.jobs.domain.JobInfo;
-import de.otto.edison.jobs.domain.JobMessage;
-import de.otto.edison.jobs.domain.RunningJobs;
+import de.otto.edison.jobs.domain.RunningJob;
 
-import java.time.OffsetDateTime;
 import java.util.List;
-import java.util.Optional;
-import java.util.Set;
 
 /**
+ * A repository that is holding locks for running or disabled jobs.
  *
  * @since 1.0.0
  */
@@ -20,24 +16,24 @@ public interface JobLockRepository {
      * blocked by some other job from the mutex group. This operation must be implemented atomically on the persistent
      * datastore (i. e. test and set) to make sure a job is never marked as running twice.
      *
-     * @param job the job to be marked
-     * @param jobTypesMutex a list of job types that must not be marked running in order to mark this job.
-     *                      The jobType to be marked will be contained in this set.
-     * @throws JobBlockedException if at least one of the jobTypes in the jobTypesMutex set is already marked running.
+     * @param jobId the id of the job
+     * @param jobType the type of the job
+     * @throws JobBlockedException if at least one of the jobTypes in the jobTypesMutex set is already marked running, or
+     * if the job type was disabled.
      */
-    void markJobAsRunningIfPossible(JobInfo job) throws JobBlockedException;
+    void aquireRunLock(String jobId, String jobType) throws JobBlockedException;
 
     /**
      * Clears the job running mark of the jobType. Does nothing if not mark exists.
      *
      * @param jobType the job type
      */
-    void clearRunningMark(String jobType);
+    void releaseRunLock(String jobType);
 
     /**
      * @return All Running Jobs as specified by the markJobAsRunningIfPossible method.
      */
-    RunningJobs runningJobs();
+    List<RunningJob> runningJobs();
 
     /**
      * Disables a job type, i.e. prevents it from being started
@@ -56,7 +52,7 @@ public interface JobLockRepository {
     /**
      * @return a list of all job types that are currently disabled
      */
-    List<String> findDisabledJobTypes();
+    List<String> disabledJobTypes();
 
     void deleteAll();