+++ edison-jobs/src/main/java/de/otto/edison/jobs/repository/JobRepository.java
@@ -2,12 +2,10 @@ package de.otto.edison.jobs.repository;
 
 import de.otto.edison.jobs.domain.JobInfo;
 import de.otto.edison.jobs.domain.JobMessage;
-import de.otto.edison.jobs.domain.RunningJobs;
 
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 public interface JobRepository {